+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `features` utilities module for checking supported features
+- Add `features` example showcasing the `features` module
+
 ### Changed
 
 - Update Vulkan Headers to version 212